NP-3537 Refactor BackgroundDiv
