+++ src/components/styled/Wrappers.tsx
@@ -1,11 +1,12 @@
+import { Box } from '@mui/material';
 import { styled } from '@mui/system';
 
-export const StyledRightAlignedWrapper = styled('div')({
+export const StyledRightAlignedWrapper = styled(Box)({
   display: 'flex',
   justifyContent: 'flex-end',
 });
 
-export const StyledSelectWrapper = styled('div')(({ theme }) => ({
+export const StyledSelectWrapper = styled(Box)(({ theme }) => ({
   width: '50%',
   maxWidth: theme.breakpoints.values.lg,
   [theme.breakpoints.down('sm')]: {
@@ -13,7 +14,7 @@ export const StyledSelectWrapper = styled('div')(({ theme }) => ({
   },
 }));
 
-export const SyledPageContent = styled('div')(({ theme }) => ({
+export const SyledPageContent = styled(Box)(({ theme }) => ({
   width: '100%',
   maxWidth: theme.breakpoints.values.lg,
   [theme.breakpoints.up('sm')]: {
@@ -22,8 +23,16 @@ export const SyledPageContent = styled('div')(({ theme }) => ({
   padding: '0.5rem',
 }));
 
-export const InputContainerBox = styled('div')({
+export const InputContainerBox = styled(Box)({
   display: 'flex',
   flexDirection: 'column',
   gap: '1.5rem',
 });
+
+export const BackgroundDiv = styled(Box)(({ theme }) => ({
+  background: theme.palette.background.paper,
+  [theme.breakpoints.down('sm')]: {
+    padding: '0.5rem',
+  },
+  padding: '1rem 2rem',
+}));
