+++ src/pages/admin/MyInstitutionPage.tsx
@@ -6,7 +6,12 @@ import { LoadingButton } from '@mui/lab';
 import { updateCustomerInstitution } from '../../api/customerInstitutionsApi';
 import { ListSkeleton } from '../../components/ListSkeleton';
 import { PageHeader } from '../../components/PageHeader';
-import { InputContainerBox, SyledPageContent, StyledRightAlignedWrapper } from '../../components/styled/Wrappers';
+import {
+  InputContainerBox,
+  SyledPageContent,
+  StyledRightAlignedWrapper,
+  BackgroundDiv,
+} from '../../components/styled/Wrappers';
 import { setNotification } from '../../redux/actions/notificationActions';
 import { RootStore } from '../../redux/reducers/rootReducer';
 import {
@@ -17,7 +22,6 @@ import {
 import { myInstitutionValidationSchema } from '../../utils/validation/customerInstitutionValidation';
 import { CustomerInstitutionTextField } from './customerInstitutionFields/CustomerInstitutionTextField';
 import { SelectInstitutionField } from './customerInstitutionFields/SelectInstitutionField';
-import { BackgroundDiv } from '../../components/BackgroundDiv';
 import { useFetch } from '../../utils/hooks/useFetch';
 import { isErrorStatus, isSuccessStatus } from '../../utils/constants';
 import { dataTestId } from '../../utils/dataTestIds';
